+++ scripts/gulpfile.ts
@@ -1,9 +1,21 @@
 import gulp from "gulp";
 import chpwd from "./change-pwd";
 import generate from "./generate";
+import genImgMap from "./generate-img-map";
+import subImg from "./substitute-img";
 
-gulp.task("chpwd", (cb) => {
-  chpwd();
+gulp.task("generate-image-map", async (cb) => {
+  await genImgMap();
+  cb();
+});
+
+gulp.task("substitute-image", (cb) => {
+  subImg();
+  cb();
+});
+
+gulp.task("change-passwd", async (cb) => {
+  await chpwd();
   cb();
 });